+++ keanu-python/keanu/context.py
@@ -1,12 +1,13 @@
-import sys
 import io
-import os
 import logging
-from py4j.java_gateway import JavaGateway, CallbackServerParameters, JavaObject, JavaClass, JVMView, java_import
-from py4j.java_collections import JavaList, JavaArray, JavaSet, JavaMap
-from py4j.protocol import Py4JError
-from typing import Dict, Any, Iterable, List, Collection, Set
+import os
+import sys
 from _io import TextIOWrapper
+from typing import Dict, Any, Iterable, Collection
+
+from py4j.java_collections import JavaList, JavaArray, JavaSet, JavaMap
+from py4j.java_gateway import JavaGateway, CallbackServerParameters, JavaObject, JavaClass, JVMView, java_import
+from py4j.protocol import Py4JError
 
 PATH = os.path.abspath(os.path.dirname(__file__))
 ND4J_CLASSPATH_ENVIRONMENT_VARIABLE = "KEANU_ND4J_CLASSPATH"
@@ -117,6 +118,9 @@ class KeanuContext(metaclass=Singleton):
 
     def to_java_int_array(self, l: Collection[Any]) -> JavaArray:
         return self.to_java_array(l, self._gateway.jvm.int)
+
+    def to_java_string_array(self, l: Collection[Any]) -> JavaArray:
+        return self.to_java_array(l, self._gateway.jvm.String)
 
     def to_java_vertex_array(self, l: Collection[Any]) -> JavaArray:
         java_import(self.jvm_view(), "io.improbable.keanu.vertices.Vertex")