+++ keanu-python/keanu/functional/function.py
@@ -9,6 +9,11 @@ class Function:
         self.lambda_function = lambda_function
 
     def apply(self, arg: JavaObject) -> JavaObject:
+        """
+        >>> f = Function(lambda x : x + 1)
+        >>> f.apply(1)
+        2
+        """
         return self.lambda_function(arg)
 
     class Java: