+++ keanu-python/keanu/vertex/base.py
@@ -38,6 +38,11 @@ class Vertex(JavaObjectWrapper, SupportsRound['Vertex']):
     def cast(self, v: tensor_arg_types) -> tensor_arg_types:
         return v
 
+    def __bool__(self) -> bool:
+        raise TypeError(
+            'Keanu vertices cannot be used as a predicate in a Python "if" statement. Please use keanu.vertex.If instead.'
+        )
+
     def __hash__(self) -> int:
         return hash(self.get_id())
 
@@ -70,6 +75,12 @@ class Vertex(JavaObjectWrapper, SupportsRound['Vertex']):
     def get_label(self) -> Optional[str]:
         label = self.unwrap().getLabel()
         return None if label is None else label.getQualifiedName()
+
+    def is_observed(self) -> bool:
+        return self.unwrap().isObserved()
+
+    def has_value(self) -> bool:
+        return self.unwrap().hasValue()
 
     """
     __array_ufunc__ is a NumPy thing that enables you to intercept and handle the numpy operation.
@@ -249,41 +260,41 @@ class Integer(Vertex):
 
     def __add__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__add__(other),
-            lambda other_holder: kn.vertex.generated.IntegerAddition(self, other_holder))
+            other, lambda casted_to_double: casted_to_double.__add__(other), lambda other_holder: kn.vertex.generated.
+            IntegerAddition(self, other_holder))
 
     def __radd__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__radd__(other),
-            lambda other_holder: kn.vertex.generated.IntegerAddition(other_holder, self))
+            other, lambda casted_to_double: casted_to_double.__radd__(other), lambda other_holder: kn.vertex.generated.
+            IntegerAddition(other_holder, self))
 
     def __sub__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__sub__(other),
-            lambda other_holder: kn.vertex.generated.IntegerDifference(self, other_holder))
+            other, lambda casted_to_double: casted_to_double.__sub__(other), lambda other_holder: kn.vertex.generated.
+            IntegerDifference(self, other_holder))
 
     def __rsub__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__rsub__(other),
-            lambda other_holder: kn.vertex.generated.IntegerDifference(other_holder, self))
+            other, lambda casted_to_double: casted_to_double.__rsub__(other), lambda other_holder: kn.vertex.generated.
+            IntegerDifference(other_holder, self))
 
     def __mul__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__mul__(other),
-            lambda other_holder: kn.vertex.generated.IntegerMultiplication(self, other_holder))
+            other, lambda casted_to_double: casted_to_double.__mul__(other), lambda other_holder: kn.vertex.generated.
+            IntegerMultiplication(self, other_holder))
 
     def __rmul__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__rmul__(other),
-            lambda other_holder: kn.vertex.generated.IntegerMultiplication(other_holder, self))
+            other, lambda casted_to_double: casted_to_double.__rmul__(other), lambda other_holder: kn.vertex.generated.
+            IntegerMultiplication(other_holder, self))
 
     def __pow__(self, other: vertex_operation_param_types) -> 'Vertex':
-        return self.__op_based_on_other_type(other, lambda casted_to_double: casted_to_double.__pow__(other),
-                                             lambda other_holder: kn.vertex.generated.IntegerPower(self, other_holder))
+        return self.__op_based_on_other_type(other, lambda casted_to_double: casted_to_double.__pow__(other), lambda
+                                             other_holder: kn.vertex.generated.IntegerPower(self, other_holder))
 
     def __rpow__(self, other: vertex_operation_param_types) -> 'Vertex':
-        return self.__op_based_on_other_type(other, lambda casted_to_double: casted_to_double.__rpow__(other),
-                                             lambda other_holder: kn.vertex.generated.IntegerPower(other_holder, self))
+        return self.__op_based_on_other_type(other, lambda casted_to_double: casted_to_double.__rpow__(other), lambda
+                                             other_holder: kn.vertex.generated.IntegerPower(other_holder, self))
 
     def __truediv__(self, other: vertex_operation_param_types) -> 'Vertex':
         return kn.vertex.generated.CastToDouble(self).__truediv__(other)
@@ -293,13 +304,13 @@ class Integer(Vertex):
 
     def __floordiv__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__truediv__(other).__floor__(),
-            lambda other_holder: kn.vertex.generated.IntegerDivision(self, other_holder))
+            other, lambda casted_to_double: casted_to_double.__truediv__(other).__floor__(), lambda other_holder: kn.
+            vertex.generated.IntegerDivision(self, other_holder))
 
     def __rfloordiv__(self, other: vertex_operation_param_types) -> 'Vertex':
         return self.__op_based_on_other_type(
-            other, lambda casted_to_double: casted_to_double.__rtruediv__(other).__floor__(),
-            lambda other_holder: kn.vertex.generated.IntegerDivision(other_holder, self))
+            other, lambda casted_to_double: casted_to_double.__rtruediv__(other).__floor__(), lambda other_holder: kn.
+            vertex.generated.IntegerDivision(other_holder, self))
 
 
 class Boolean(Vertex):